redesign architecture

+++ SuddenAttack/source/Script/Account.h
@@ -10,8 +10,8 @@ public:
 	Account(ze::GameObject& owner);
 	virtual ~Account() = default;
 
-	uint64_t GetAccountId() const { return m_accountId; }
-	void SetAccountId(uint64_t id) { m_accountId = id; }
+	uint32_t GetAccountId() const { return m_accountId; }
+	void SetAccountId(uint32_t id) { m_accountId = id; }
 	uint16_t GetLevel() const { return m_level; }
 	void SetLevel(uint16_t level) { m_level = level; }
 	uint32_t GetExp() const { return m_exp; }
@@ -22,7 +22,7 @@ public:
 	const wchar_t* GetNickname() const { return m_nickname; }
 	void SetNickname(const wchar_t* nickname, uint16_t len);
 private:
-	uint64_t m_accountId;
+	uint32_t m_accountId;
 	uint16_t m_level;	// 레벨
 	uint32_t m_exp;		// 경험치
 	uint32_t m_point;	// 포인트 소유량
